+++ src/videojs-http-streaming.js
@@ -630,6 +630,7 @@ class VhsHandler extends Component {
       typeof this.source_.useBandwidthFromLocalStorage !== 'undefined' ?
         this.source_.useBandwidthFromLocalStorage :
         this.options_.useBandwidthFromLocalStorage || false;
+    this.options_.useNetworkInformationApi = this.options_.useNetworkInformationApi || false;
     this.options_.customTagParsers = this.options_.customTagParsers || [];
     this.options_.customTagMappers = this.options_.customTagMappers || [];
     this.options_.cacheEncryptionKeys = this.options_.cacheEncryptionKeys || false;
@@ -682,6 +683,7 @@ class VhsHandler extends Component {
       'experimentalBufferBasedABR',
       'liveRangeSafeTimeDelta',
       'experimentalLLHLS',
+      'useNetworkInformationApi',
       'experimentalExactManifestTimings',
       'experimentalLeastPixelDiffSelector'
     ].forEach((option) => {
@@ -789,7 +791,27 @@ class VhsHandler extends Component {
       },
       bandwidth: {
         get() {
-          return this.masterPlaylistController_.mainSegmentLoader_.bandwidth;
+          let playerBandwidthEst = this.masterPlaylistController_.mainSegmentLoader_.bandwidth;
+
+          const networkInformation = window.navigator.connection || window.navigator.mozConnection || window.navigator.webkitConnection;
+          const tenMbpsAsBitsPerSecond = 10e6;
+
+          if (this.options_.useNetworkInformationApi && networkInformation) {
+            // downlink returns Mbps
+            // https://developer.mozilla.org/en-US/docs/Web/API/NetworkInformation/downlink
+            const networkInfoBandwidthEstBitsPerSec = networkInformation.downlink * 1000 * 1000;
+
+            // downlink maxes out at 10 Mbps. In the event that both networkInformationApi and the player
+            // estimate a bandwidth greater than 10 Mbps, use the larger of the two estimates to ensure that
+            // high quality streams are not filtered out.
+            if (networkInfoBandwidthEstBitsPerSec >= tenMbpsAsBitsPerSecond && playerBandwidthEst >= tenMbpsAsBitsPerSecond) {
+              playerBandwidthEst = Math.max(playerBandwidthEst, networkInfoBandwidthEstBitsPerSec);
+            } else {
+              playerBandwidthEst = networkInfoBandwidthEstBitsPerSec;
+            }
+          }
+
+          return playerBandwidthEst;
         },
         set(bandwidth) {
           this.masterPlaylistController_.mainSegmentLoader_.bandwidth = bandwidth;
@@ -982,6 +1004,41 @@ class VhsHandler extends Component {
     this.tech_.src(this.mediaSourceUrl_);
   }
 
+  createKeySessions_() {
+    const audioPlaylistLoader =
+      this.masterPlaylistController_.mediaTypes_.AUDIO.activePlaylistLoader;
+
+    this.logger_('waiting for EME key session creation');
+    waitForKeySessionCreation({
+      player: this.player_,
+      sourceKeySystems: this.source_.keySystems,
+      audioMedia: audioPlaylistLoader && audioPlaylistLoader.media(),
+      mainPlaylists: this.playlists.master.playlists
+    }).then(() => {
+      this.logger_('created EME key session');
+      this.masterPlaylistController_.sourceUpdater_.initializedEme();
+    }).catch((err) => {
+      this.logger_('error while creating EME key session', err);
+      this.player_.error({
+        message: 'Failed to initialize media keys for EME',
+        code: 3
+      });
+    });
+  }
+
+  handleWaitingForKey_() {
+    // If waitingforkey is fired, it's possible that the data that's necessary to retrieve
+    // the key is in the manifest. While this should've happened on initial source load, it
+    // may happen again in live streams where the keys change, and the manifest info
+    // reflects the update.
+    //
+    // Because videojs-contrib-eme compares the PSSH data we send to that of PSSH data it's
+    // already requested keys for, we don't have to worry about this generating extraneous
+    // requests.
+    this.logger_('waitingforkey fired, attempting to create any new key sessions');
+    this.createKeySessions_();
+  }
+
   /**
    * If necessary and EME is available, sets up EME options and waits for key session
    * creation.
@@ -1011,6 +1068,9 @@ class VhsHandler extends Component {
       }
     });
 
+    this.handleWaitingForKey_ = this.handleWaitingForKey_.bind(this);
+    this.player_.tech_.on('waitingforkey', this.handleWaitingForKey_);
+
     // In IE11 this is too early to initialize media keys, and IE11 does not support
     // promises.
     if (videojs.browser.IE_VERSION === 11 || !didSetupEmeOptions) {
@@ -1019,22 +1079,7 @@ class VhsHandler extends Component {
       return;
     }
 
-    this.logger_('waiting for EME key session creation');
-    waitForKeySessionCreation({
-      player: this.player_,
-      sourceKeySystems: this.source_.keySystems,
-      audioMedia: audioPlaylistLoader && audioPlaylistLoader.media(),
-      mainPlaylists: this.playlists.master.playlists
-    }).then(() => {
-      this.logger_('created EME key session');
-      this.masterPlaylistController_.sourceUpdater_.initializedEme();
-    }).catch((err) => {
-      this.logger_('error while creating EME key session', err);
-      this.player_.error({
-        message: 'Failed to initialize media keys for EME',
-        code: 3
-      });
-    });
+    this.createKeySessions_();
   }
 
   /**
@@ -1149,6 +1194,10 @@ class VhsHandler extends Component {
       this.mediaSourceUrl_ = null;
     }
 
+    if (this.tech_) {
+      this.tech_.off('waitingforkey', this.handleWaitingForKey_);
+    }
+
     super.dispose();
   }
 
@@ -1208,10 +1257,14 @@ const VhsSourceHandler = {
     return tech.vhs;
   },
   canPlayType(type, options = {}) {
-    const { vhs: { overrideNative = !videojs.browser.IS_ANY_SAFARI } } = videojs.mergeOptions(videojs.options, options);
+    const {
+      vhs: { overrideNative = !videojs.browser.IS_ANY_SAFARI } = {},
+      hls: { overrideNative: legacyOverrideNative = false } = {}
+    } = videojs.mergeOptions(videojs.options, options);
+
     const supportedType = simpleTypeFromSourceType(type);
     const canUseMsePlayback = supportedType &&
-      (!Vhs.supportsTypeNatively(supportedType) || overrideNative);
+      (!Vhs.supportsTypeNatively(supportedType) || legacyOverrideNative || overrideNative);
 
     return canUseMsePlayback ? 'maybe' : '';
   }