+++ test/videojs-http-streaming.test.js
@@ -416,7 +416,7 @@ QUnit.test('logs deprecation notice when using hls for options', function(assert
   assert.equal(this.env.log.warn.calls, 1, 'warning logged');
   assert.equal(
     this.env.log.warn.args[0][0],
-    'Using hls options is deprecated. Use vhs instead.',
+    'Using hls options is deprecated. Use vhs instead.'
   );
 });
 
@@ -440,7 +440,7 @@ QUnit.test('logs deprecation notice when using hls for global options', function
   assert.equal(this.env.log.warn.calls, 1, 'warning logged');
   assert.equal(
     this.env.log.warn.args[0][0],
-    'Using hls options is deprecated. Use vhs instead.',
+    'Using hls options is deprecated. Use vhs instead.'
   );
 
   videojs.options.hls = origHlsOptions;
@@ -1159,6 +1159,120 @@ QUnit.test(
     );
   }
 );
+
+QUnit.module('NetworkInformationApi', hooks => {
+  hooks.beforeEach(function(assert) {
+    this.env = useFakeEnvironment(assert);
+    this.ogNavigator = window.navigator;
+    this.clock = this.env.clock;
+
+    this.resetNavigatorConnection = (connection = {}) => {
+      // Need to delete the property before setting since navigator doesn't have a setter
+      delete window.navigator;
+      window.navigator = {
+        connection
+      };
+    };
+  });
+
+  hooks.afterEach(function() {
+    this.env.restore();
+    window.navigator = this.ogNavigator;
+  });
+
+  QUnit[testOrSkip](
+    'bandwidth returns networkInformation.downlink when useNetworkInformationApi option is enabled',
+    function(assert) {
+      this.resetNavigatorConnection({
+        downlink: 10
+      });
+      this.player = createPlayer({ html5: { vhs: { useNetworkInformationApi: true } } });
+      this.player.src({
+        src: 'manifest/master.m3u8',
+        type: 'application/vnd.apple.mpegurl'
+      });
+
+      this.clock.tick(1);
+
+      // downlink in bits = 10 * 1000000 = 10e6
+      assert.strictEqual(
+        this.player.tech_.vhs.bandwidth,
+        10e6,
+        'bandwidth equals networkInfo.downlink represented as bits per second'
+      );
+    }
+  );
+
+  QUnit[testOrSkip](
+    'bandwidth uses player-estimated bandwidth when its value is greater than networkInformation.downLink and both values are >= 10 Mbps',
+    function(assert) {
+      this.resetNavigatorConnection({
+        // 10 Mbps or 10e6
+        downlink: 10
+      });
+      this.player = createPlayer({ html5: { vhs: { useNetworkInformationApi: true } } });
+      this.player.src({
+        src: 'manifest/master.m3u8',
+        type: 'application/vnd.apple.mpegurl'
+      });
+
+      this.clock.tick(1);
+
+      this.player.tech_.vhs.bandwidth = 20e6;
+      assert.strictEqual(
+        this.player.tech_.vhs.bandwidth,
+        20e6,
+        'bandwidth getter returned the player-estimated bandwidth value'
+      );
+    }
+  );
+
+  QUnit[testOrSkip](
+    'bandwidth uses network-information-api bandwidth when its value is less than the player bandwidth and 10 Mbps',
+    function(assert) {
+      this.resetNavigatorConnection({
+        // 9 Mbps or 9e6
+        downlink: 9
+      });
+      this.player = createPlayer({ html5: { vhs: { useNetworkInformationApi: true } } });
+      this.player.src({
+        src: 'manifest/master.m3u8',
+        type: 'application/vnd.apple.mpegurl'
+      });
+
+      this.clock.tick(1);
+
+      this.player.tech_.vhs.bandwidth = 20e10;
+      assert.strictEqual(
+        this.player.tech_.vhs.bandwidth,
+        9e6,
+        'bandwidth getter returned the network-information-api bandwidth value since it was less than 10 Mbps'
+      );
+    }
+  );
+
+  QUnit[testOrSkip](
+    'bandwidth uses player-estimated bandwidth when networkInformation is not supported',
+    function(assert) {
+      // Nullify the `connection` property on Navigator
+      this.resetNavigatorConnection(null);
+      this.player = createPlayer({ html5: { vhs: { useNetworkInformationApi: true } } });
+      this.player.src({
+        src: 'manifest/master.m3u8',
+        type: 'application/vnd.apple.mpegurl'
+      });
+
+      this.clock.tick(1);
+
+      this.player.tech_.vhs.bandwidth = 20e10;
+      assert.strictEqual(
+        this.player.tech_.vhs.bandwidth,
+        20e10,
+        'bandwidth getter returned the player-estimated bandwidth value'
+      );
+    }
+  );
+});
 
 QUnit.test('requests a reasonable rendition to start', function(assert) {
   this.player.src({
@@ -4599,6 +4713,55 @@ QUnit.test('eme handles keystatuschange where status is usable', function(assert
   assert.deepEqual(excludes, [], 'did not exclude anything');
 });
 
+QUnit.test('eme waitingforkey event triggers another setup', function(assert) {
+  this.player.eme = { options: { setting: 1 } };
+  this.player.src({
+    src: 'manifest/master.m3u8',
+    type: 'application/x-mpegURL',
+    keySystems: { keySystem1: { url: 'url1' } }
+  });
+
+  this.clock.tick(1);
+
+  const media = {
+    attributes: { CODECS: 'avc1.420015, mp4a.40.2c' },
+    contentProtection: { keySystem1: { pssh: 'test' } }
+  };
+
+  const vhs = this.player.tech_.vhs;
+
+  vhs.playlists = {
+    master: { playlists: [media] },
+    media: () => media
+  };
+
+  const origCreateKeySessions = vhs.createKeySessions_.bind(vhs);
+  let createKeySessionCalls = 0;
+
+  vhs.createKeySessions_ = () => {
+    createKeySessionCalls++;
+    origCreateKeySessions();
+  };
+
+  vhs.masterPlaylistController_.sourceUpdater_.trigger('createdsourcebuffers');
+
+  // Since IE11 doesn't initialize media keys early, in this test IE11 will always have
+  // one less call than in other browsers.
+  if (videojs.browser.IE_VERSION === 11) {
+    assert.equal(createKeySessionCalls, 0, 'did not call createKeySessions_ yet');
+  } else {
+    assert.equal(createKeySessionCalls, 1, 'called createKeySessions_ once');
+  }
+
+  this.player.tech_.trigger({type: 'waitingforkey', status: 'usable'});
+
+  if (videojs.browser.IE_VERSION === 11) {
+    assert.equal(createKeySessionCalls, 1, 'called createKeySessions_ once');
+  } else {
+    assert.equal(createKeySessionCalls, 2, 'called createKeySessions_ again');
+  }
+});
+
 QUnit.test('integration: configures eme for DASH on source buffer creation', function(assert) {
   assert.timeout(3000);
   const done = assert.async();
@@ -5880,7 +6043,7 @@ QUnit.test('emeKeySystems adds content types for all keySystems', function(asser
   assert.deepEqual(
     emeKeySystems(
       { keySystem1: {}, keySystem2: {} },
-      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } },
+      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } }
     ),
     {
       keySystem1: {
@@ -5900,7 +6063,7 @@ QUnit.test('emeKeySystems adds content types for all keySystems', function(asser
     emeKeySystems(
       { keySystem1: {}, keySystem2: {} },
       { attributes: { CODECS: 'avc1.420015' } },
-      { attributes: { CODECS: 'mp4a.40.2c' } },
+      { attributes: { CODECS: 'mp4a.40.2c' } }
     ),
     {
       keySystem1: {
@@ -5920,7 +6083,7 @@ QUnit.test('emeKeySystems supports audio only', function(assert) {
   assert.deepEqual(
     emeKeySystems(
       { keySystem1: {}, keySystem2: {} },
-      { attributes: { CODECS: 'mp4a.40.2c' } },
+      { attributes: { CODECS: 'mp4a.40.2c' } }
     ),
     {
       keySystem1: {
@@ -5939,7 +6102,7 @@ QUnit.test('emeKeySystems supports external audio only', function(assert) {
     emeKeySystems(
       { keySystem1: {}, keySystem2: {} },
       { attributes: {} },
-      { attributes: { CODECS: 'mp4a.40.2c' } },
+      { attributes: { CODECS: 'mp4a.40.2c' } }
     ),
     {
       keySystem1: {
@@ -5957,7 +6120,7 @@ QUnit.test('emeKeySystems supports video only', function(assert) {
   assert.deepEqual(
     emeKeySystems(
       { keySystem1: {}, keySystem2: {} },
-      { attributes: { CODECS: 'avc1.420015' } },
+      { attributes: { CODECS: 'avc1.420015' } }
     ),
     {
       keySystem1: {
@@ -5975,7 +6138,7 @@ QUnit.test('emeKeySystems retains non content type properties', function(assert)
   assert.deepEqual(
     emeKeySystems(
       { keySystem1: { url: '1' }, keySystem2: { url: '2'} },
-      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } },
+      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } }
     ),
     {
       keySystem1: {
@@ -6006,7 +6169,7 @@ QUnit.test('emeKeySystems overwrites content types', function(assert) {
           videoContentType: 'd'
         }
       },
-      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } },
+      { attributes: { CODECS: 'avc1.420015, mp4a.40.2c' } }
     ),
     {
       keySystem1: {